moved keycloak to src folder
