+++ backend/src/keycloak/activateUser.ts
